MEALMentor/pyproject.toml
[project]
name = "mealmentor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
